Move encryption to its own module
