+++ client/go.mod
@@ -1,0 +1,3 @@
+module client
+
+go 1.23.2
