+++ security/security.go
@@ -1,4 +1,4 @@
-package client
+package security
 
 import (
 	"crypto/rand"
